requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
colorama==0.4.6
Django==4.1.4
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
msgpack==1.0.4
numpy==1.24.1
openai==0.25.0
opencv-python-headless==4.7.0.68
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7
redis==4.5.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
wkhtmltopdf==0.2
